Alteração do relatório para conexão com o Ws Sefaz

+++ Escon.SisctNET.Model/DarDocument.cs
@@ -15,6 +15,6 @@ namespace Escon.SisctNET.Model
 
         public int ControlNumber { get; set; }
 
-        public int DocumentNumber { get; set; }
+        public long DocumentNumber { get; set; }
     }
 }
